Move assistant replay tracker out of loop
